This function returns a comprehensive list of English stopwords that allows adding more customized words.

This function returns a comprehensive list of English stopwords that allows adding more customized words.

CleverStopwords(words)

Arguments

words

A list of strings that needs to be included in the stopwords.

Value

stopwords_all A list of strings that includes the complete set of stopwords.

Examples

words <- list("would", "aaa")
CleverStopwords(words)
#> [[1]]
#> [1] "a"
#> 
#> [[2]]
#> [1] "aaa"
#> 
#> [[3]]
#> [1] "about"
#> 
#> [[4]]
#> [1] "above"
#> 
#> [[5]]
#> [1] "after"
#> 
#> [[6]]
#> [1] "again"
#> 
#> [[7]]
#> [1] "against"
#> 
#> [[8]]
#> [1] "all"
#> 
#> [[9]]
#> [1] "am"
#> 
#> [[10]]
#> [1] "an"
#> 
#> [[11]]
#> [1] "and"
#> 
#> [[12]]
#> [1] "any"
#> 
#> [[13]]
#> [1] "are"
#> 
#> [[14]]
#> [1] "aren't"
#> 
#> [[15]]
#> [1] "as"
#> 
#> [[16]]
#> [1] "at"
#> 
#> [[17]]
#> [1] "be"
#> 
#> [[18]]
#> [1] "because"
#> 
#> [[19]]
#> [1] "been"
#> 
#> [[20]]
#> [1] "before"
#> 
#> [[21]]
#> [1] "being"
#> 
#> [[22]]
#> [1] "below"
#> 
#> [[23]]
#> [1] "between"
#> 
#> [[24]]
#> [1] "both"
#> 
#> [[25]]
#> [1] "but"
#> 
#> [[26]]
#> [1] "by"
#> 
#> [[27]]
#> [1] "can't"
#> 
#> [[28]]
#> [1] "cannot"
#> 
#> [[29]]
#> [1] "could"
#> 
#> [[30]]
#> [1] "couldn't"
#> 
#> [[31]]
#> [1] "did"
#> 
#> [[32]]
#> [1] "didn't"
#> 
#> [[33]]
#> [1] "do"
#> 
#> [[34]]
#> [1] "does"
#> 
#> [[35]]
#> [1] "doesn't"
#> 
#> [[36]]
#> [1] "doing"
#> 
#> [[37]]
#> [1] "don't"
#> 
#> [[38]]
#> [1] "down"
#> 
#> [[39]]
#> [1] "during"
#> 
#> [[40]]
#> [1] "each"
#> 
#> [[41]]
#> [1] "few"
#> 
#> [[42]]
#> [1] "for"
#> 
#> [[43]]
#> [1] "from"
#> 
#> [[44]]
#> [1] "further"
#> 
#> [[45]]
#> [1] "had"
#> 
#> [[46]]
#> [1] "hadn't"
#> 
#> [[47]]
#> [1] "has"
#> 
#> [[48]]
#> [1] "hasn't"
#> 
#> [[49]]
#> [1] "have"
#> 
#> [[50]]
#> [1] "haven't"
#> 
#> [[51]]
#> [1] "having"
#> 
#> [[52]]
#> [1] "he"
#> 
#> [[53]]
#> [1] "he'd"
#> 
#> [[54]]
#> [1] "he'll"
#> 
#> [[55]]
#> [1] "he's"
#> 
#> [[56]]
#> [1] "her"
#> 
#> [[57]]
#> [1] "here"
#> 
#> [[58]]
#> [1] "here's"
#> 
#> [[59]]
#> [1] "hers"
#> 
#> [[60]]
#> [1] "herself"
#> 
#> [[61]]
#> [1] "him"
#> 
#> [[62]]
#> [1] "himself"
#> 
#> [[63]]
#> [1] "his"
#> 
#> [[64]]
#> [1] "how"
#> 
#> [[65]]
#> [1] "how's"
#> 
#> [[66]]
#> [1] "i"
#> 
#> [[67]]
#> [1] "i'd"
#> 
#> [[68]]
#> [1] "i'll"
#> 
#> [[69]]
#> [1] "i'm"
#> 
#> [[70]]
#> [1] "i've"
#> 
#> [[71]]
#> [1] "if"
#> 
#> [[72]]
#> [1] "in"
#> 
#> [[73]]
#> [1] "into"
#> 
#> [[74]]
#> [1] "is"
#> 
#> [[75]]
#> [1] "isn't"
#> 
#> [[76]]
#> [1] "it"
#> 
#> [[77]]
#> [1] "it's"
#> 
#> [[78]]
#> [1] "its"
#> 
#> [[79]]
#> [1] "itself"
#> 
#> [[80]]
#> [1] "let's"
#> 
#> [[81]]
#> [1] "me"
#> 
#> [[82]]
#> [1] "more"
#> 
#> [[83]]
#> [1] "most"
#> 
#> [[84]]
#> [1] "mustn't"
#> 
#> [[85]]
#> [1] "my"
#> 
#> [[86]]
#> [1] "myself"
#> 
#> [[87]]
#> [1] "no"
#> 
#> [[88]]
#> [1] "nor"
#> 
#> [[89]]
#> [1] "not"
#> 
#> [[90]]
#> [1] "of"
#> 
#> [[91]]
#> [1] "off"
#> 
#> [[92]]
#> [1] "on"
#> 
#> [[93]]
#> [1] "once"
#> 
#> [[94]]
#> [1] "only"
#> 
#> [[95]]
#> [1] "or"
#> 
#> [[96]]
#> [1] "other"
#> 
#> [[97]]
#> [1] "ought"
#> 
#> [[98]]
#> [1] "our"
#> 
#> [[99]]
#> [1] "ours"
#> 
#> [[100]]
#> [1] "ourselves"
#> 
#> [[101]]
#> [1] "out"
#> 
#> [[102]]
#> [1] "over"
#> 
#> [[103]]
#> [1] "own"
#> 
#> [[104]]
#> [1] "same"
#> 
#> [[105]]
#> [1] "shan't"
#> 
#> [[106]]
#> [1] "she"
#> 
#> [[107]]
#> [1] "she'd"
#> 
#> [[108]]
#> [1] "she'll"
#> 
#> [[109]]
#> [1] "she's"
#> 
#> [[110]]
#> [1] "should"
#> 
#> [[111]]
#> [1] "shouldn't"
#> 
#> [[112]]
#> [1] "so"
#> 
#> [[113]]
#> [1] "some"
#> 
#> [[114]]
#> [1] "such"
#> 
#> [[115]]
#> [1] "than"
#> 
#> [[116]]
#> [1] "that"
#> 
#> [[117]]
#> [1] "that's"
#> 
#> [[118]]
#> [1] "the"
#> 
#> [[119]]
#> [1] "their"
#> 
#> [[120]]
#> [1] "theirs"
#> 
#> [[121]]
#> [1] "them"
#> 
#> [[122]]
#> [1] "themselves"
#> 
#> [[123]]
#> [1] "then"
#> 
#> [[124]]
#> [1] "there"
#> 
#> [[125]]
#> [1] "there's"
#> 
#> [[126]]
#> [1] "these"
#> 
#> [[127]]
#> [1] "they"
#> 
#> [[128]]
#> [1] "they'd"
#> 
#> [[129]]
#> [1] "they'll"
#> 
#> [[130]]
#> [1] "they're"
#> 
#> [[131]]
#> [1] "they've"
#> 
#> [[132]]
#> [1] "this"
#> 
#> [[133]]
#> [1] "those"
#> 
#> [[134]]
#> [1] "through"
#> 
#> [[135]]
#> [1] "to"
#> 
#> [[136]]
#> [1] "too"
#> 
#> [[137]]
#> [1] "under"
#> 
#> [[138]]
#> [1] "until"
#> 
#> [[139]]
#> [1] "up"
#> 
#> [[140]]
#> [1] "very"
#> 
#> [[141]]
#> [1] "was"
#> 
#> [[142]]
#> [1] "wasn't"
#> 
#> [[143]]
#> [1] "we"
#> 
#> [[144]]
#> [1] "we'd"
#> 
#> [[145]]
#> [1] "we'll"
#> 
#> [[146]]
#> [1] "we're"
#> 
#> [[147]]
#> [1] "we've"
#> 
#> [[148]]
#> [1] "were"
#> 
#> [[149]]
#> [1] "weren't"
#> 
#> [[150]]
#> [1] "what"
#> 
#> [[151]]
#> [1] "what's"
#> 
#> [[152]]
#> [1] "when"
#> 
#> [[153]]
#> [1] "when's"
#> 
#> [[154]]
#> [1] "where"
#> 
#> [[155]]
#> [1] "where's"
#> 
#> [[156]]
#> [1] "which"
#> 
#> [[157]]
#> [1] "while"
#> 
#> [[158]]
#> [1] "who"
#> 
#> [[159]]
#> [1] "who's"
#> 
#> [[160]]
#> [1] "whom"
#> 
#> [[161]]
#> [1] "why"
#> 
#> [[162]]
#> [1] "why's"
#> 
#> [[163]]
#> [1] "will"
#> 
#> [[164]]
#> [1] "with"
#> 
#> [[165]]
#> [1] "won't"
#> 
#> [[166]]
#> [1] "would"
#> 
#> [[167]]
#> [1] "wouldn't"
#> 
#> [[168]]
#> [1] "you"
#> 
#> [[169]]
#> [1] "you'd"
#> 
#> [[170]]
#> [1] "you'll"
#> 
#> [[171]]
#> [1] "you're"
#> 
#> [[172]]
#> [1] "you've"
#> 
#> [[173]]
#> [1] "your"
#> 
#> [[174]]
#> [1] "yours"
#> 
#> [[175]]
#> [1] "yourself"
#> 
#> [[176]]
#> [1] "yourselves"
#>